requirements.txt
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
openpyxl==3.1.2
Pillow>=10.2.0,<11.0.0